+++ src/pages/archived-conversations/archived-conversations.ts
@@ -88,7 +88,9 @@ export class ArchivedConversationsPage {
   openMessageList(archivedConversation) {
     const that = this;
     setTimeout(function () {
+      const archivedConversation = that.archivedConversations.find(item => item.uid === that.uidConvSelected);
       if (archivedConversation) {
+      //if (archivedConversation) {
         archivedConversation.is_new = false;
         archivedConversation.status = '0';
         archivedConversation.selected = true;
@@ -98,10 +100,11 @@ export class ArchivedConversationsPage {
           channel_type: archivedConversation.channel_type
         });
         that.databaseProvider.setUidLastOpenConversation(archivedConversation.uid);
-        that.events.publish('uidConvSelected:changed', archivedConversation.uid);
+        that.events.publish('uidConvSelected:changed', archivedConversation.uid, 'archived');
         that.conversationSelected = archivedConversation;
       }
     }, 0);
   }
+
 
 }